use traditional R source command when magic comment missing; fixes #3847
This fixes an exception when saving a file after deleting the !source
magic comment. When this happens, the file is still marked with the
custom source extended type (as it hasn't been saved yet), but doesn't
actually have the magic comment. The exception occurs when we try to
read the just-deleted comment to form the source command.