Persist visibility of Jobs tab
If you close the Jobs tab, the next time you start the IDE the tab will not be shown. This is same model as the Terminal Tab. Allows a user who never uses the feature to hide it once and not see it again.

To get Jobs tab back, use "View/Show Jobs" or run a script as a job via editor drop down "Source / Source as Job" or "Tools/Jobs/Start Job".

Fixes #3743