systemd: Require network-online.target in the [Unit] section
At my organization, during a reboot of our RStudio host, we encountered
a race condition where the server would not be able to resolve hosts
defined as load balancing targets, and would therefore fail to start.

Can't RStudio just be configured to use network-online.target, and
refuse to start if that is not the case? I could see potential issues
with e.g. running on a system that is not online, but at the same time
it seems like this makes a lot more sense.

This CL applies After= and Wants=network-online.target.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>