Print Messages and Warnings differently than Errors
This is a simple proof of concept to address https://github.com/rstudio/rstudio/issues/2574

There are some open issues it would be nice to discuss.

Warnings seem to consist of 2 separate error calls resulting in two
separate span blocks. The first is just the text 'Warning:' the second
is the actual warning message. This causes the current code to only color the
text of 'Warning:', which I think actually looks pretty nice, maybe we
could do the same for errors?

Errors (at least short ones) have the entire error in one span, so the
whole thing is colored.

The search string assumes an English locale, it would likely have to be
localized, as the 'Error' text changes when other languages are used.

This only changes the behavior in the R shell, it seems
getThemeErrorClass() is called a few more places, they could possibly
need to be changed as well.

We would probably also need to change consoleWriteExtendedError, but I
was unsure when this is called so did not know how to test it.

Fixes https://github.com/rstudio/rstudio/issues/2574