Fix TerminalTab's deferred loading
TerminalTab was always triggering load of the presenter due to implementing ProvidesBusy, which triggers a call to AddBusyHandler on the shim; fixed by removing this as the TerminalTab doesn't use this mechanism for reporting its busy state (it was going to at one point very early on, but ended up with its own busy-tracking mechanism).

Also fixed a typo and removed an unused initialize() method in RenderRmdOutputTab and Presenter.