fix dockerfile builds
Several builds were failing as a result of clang being built unnecessarily, both on platforms which already have clang 4+, and on platforms where we don't need it.

This was caused when I started running install-crashpad from install-common via sudo, which is needed when setting up a dev environment but not a good idea in Docker; caused path and environment variables not to be seen.

Fixed to make default behavior to not run via sudo and add opt-in flag to install-common for use by dependency bootstrapper scripts.