work around GWT devmode bug involving method references
Using a method reference causes an issue where gwt devmode throws away WizardPageSelector class, so going back to a lambda.

This manifested when I touched WizardPageSelector.java and tried to reload the page and use a wizard. It failed and even reverting the file wouldn't fix it (some order-of-operation thing in the GWT compiler, it appears).