add unique IDs to New Shiny Web App dialog box
- Name textbox: `rstudio_new_shiny_app_name`
- Single file radio: `rstudio_new_shiny_app_single_file`
- Multi file radio: `rstudio_new_shiny_app_multi_file`
- Create within directory textbox: `rstudio_textboxbutton_text`
- Browse button: `rstudio_textboxbutton_button`
- Create button: `rstudio_dlg_ok`
- Cancel button: `rstudio_dlg_cancel`

Radio buttons are compound controls. The unique ID is attached to the outer span which contains the input (radio) and label.

The create and cancel buttons already had unique ids, just including here for completeness.

The read-only textfield and browse button are part of a compound control, TextBoxWithButton, used in many places. If multiple instances of this control are simultaneously displayed (e.g. in the same dialog or stacked dialogs) then the IDs will be kept unique by addition of "_##" where ## is an incrementing integer. These IDs would be stable for a given scenario, so still usable for automation.