use common compilation args when generating precompiled headers
This is necessary as, otherwise, the precompiled headers may effectively
be incompatible with the to-be-compiled translation units used in the
package.

As an example, if the precompiled headers are generated without
'-fopenmp', but C++ files within the package are compiled with
'-fopenmp', then attempts to include precompiled headers will fail and
diagnostics will not work as expected.

Part of #4685.