When screen reader support enabled, reduce UI animations
The UI animations are not useful or necessary to a screen-reader user and may confuse screen reader software, so enable the "reduce motion" setting automatically when screen-reader is enabled, and turn it off when screen reader is disabled.

Potentially annoying to a user who frequently switches between screen-reader on / off and prefers no animations in both cases; they will have to manually turn animations back off after disabling screen reader. But fancier handling of this (expected to be) very rare case specially seems like overkill.