simplify screen reader toggle shortcut
Command that enables/disabled screen reader support had a 4-finger shortcut (Ctrl+Alt+Shift+/). Any shortcut requiring 4 fingers is unpleasant, but especially one that is so integral to bootstrapping the use of a screen reader in RStudio.

Changed to Alt+Shift+/ (but also left in the other mapping since it's been published on the blog). Will go back and update the blog to just reference the shorter one.