simplify several accessibility-related shortcuts
Change the letter used to access the Accessibility submenu to "A" and the letter used for About to "B". Keyboard users will be more likely to use the Accessibility submenu than the About dialog, so having the most memorable key be used is ideal.

Based on user feedback, changing several accessibility keyboard shortcuts to be 3-finger instead of 4-finger:

- Toggle Tab Key Moves Focus is now Alt+Shift+[ on Windows and Linux, and Ctrl+Option+[ on Mac
- Speak Editor Location is  now Alt+Shift+1 on Windows and Linux, and Ctrl+Option+1 on Mac
- These are the last planned accessibility-related shortcut tweaks for 1.3 (barring discovery of conflicts)