On Linux, do not bundle libpq and libsqlite3.
This is generally a bad idea if it can be avoided, and it can on Linux.
We will now simply rely on the system package manager to install these
dependencies, as they are very common and present in all of the major
official repos.

This as the additional benefit of being less error prone, as before
we were not also bundling all possible dependencies of the client
libraries (like libldap), which could cause a failure to load the
executables if those were not previously installed.