Prefer static linking of SOCI.
This change makes it so that if available, we link to the static
SOCI libraries instead of the dynamic ones. Many RStudio binaries
cannot use the RPATH technique of finding the path to our shipped
SOCI shared objects, which necessitates the use of static linking.
This is because many binaries are setuid binaries, and their child
processes are explicitly forbidden by the loader to use RPATH,
thus preventing the child processes from starting.

We continue to provide a means for linking to the system SOCI shared
libraries if no static libraries are present, allowing external
package maintainers a little flexibility when building RStudio.