allow enter key to click active button in web dialogs (#6952)
Currently, the Enter key clicks the default button in web dialogs (which can be different than the button with keyboard focus). This matches behavior of MacOS desktop dialogs, and (I think) dialogs in Windows many versions ago, but not current Windows desktop apps and not web applications. Per aria, focused button should be clickable with spacebar or enter key. I confirmed this is common practice in other web applications such as Google Docs (e.g. the File / Page Setup dialog).

Change allows buttons in RStudio web dialogs to handle the enter key instead the dialog itself. The "default" behavior still kicks in when focus is on other controls, but buttons are now clickable with spacebar or enter key.

Fixes #6928