Fix perpetually busy Git-Bash terminal on Windows (#7123)
* Fix perpetually busy Git-Bash terminal on Windows

- Fixes #7112

The Twisted Tale:

When I fixed how we launch Git-Bash in #6828 that gave us one "bash.exe" launching a second "bash.exe" as a child process (which was intentional). This is then caught by the subprocess-detection and flagged as always busy. Previously, git-bash was a single process. How I never noticed this I'm not sure.

The crux of this PR is that we never intended to support subprocess detection for Terminals on Windows Desktop, so fully disabling it. It's never been possible for git-bash or WSL because commands you run in those environments aren't actually Win32 processes, so can't be detected. It does work (somewhat) on Command Prompt or PowerShell, but the intention was to simply be consistent and not support it. Now that is actually the case.

I've made this change in the safest way I can think of, and it is completely scoped to Windows Desktop-only.

Note that as part of "not supporting it" we've never shown "Ask before killing processes" setting in Global Options on Windows (which would have been a workaround for this if we did).

* fix unit test