Better handle last minute DOI discovery
If we are handling a paste and are about to show the insert cite dialog and we discover that the bibilography already contains a DOI (this moment may be the first time we’ve been guaranteed a completely loaded bibliography), we should just write the cite id. It isn’t a perfect experience, but should be an edge case as bibliographies are typically loaded.