no longer update selection on 'change' (handled by 'changeSelection')
This was required to fix an out of range exception that occurred when backspacing over the entire contents of an rmd chunk. This was likely here previously as partial compensation for the 'changeCursor' event not firing (so it made some things work that didn't). Now that we have a working 'changeSelection' we can safely remove this to fix the out of range error.