Extend package description
I don't expect this to be the final commit here, but I noticed this description when I was installing from .deb just now and it struck me as being maybe a bit stale. The blame says this hasn't changed in 10 years so that kind of confirmed my suspicions.

This PR is mainly to draw attention to this, and offer a suggested improvement, but feel free to make further edits here as you see fit.