Filter out items that are moved to the trash
For items in the trash, Zotero adds an entry in the deletedItems table. Use this table to filter out items that were placed in the trash. (When the trash it emptied, the items are actually deleted).

fixes rstudio/rstudio#7722

Note that when Zotero moves an item to the trash, it doesn’t increment the item (and therefore library) version. This means that the server cache may retain that item as outside the trash until the cache is refreshed by anything that increments the library version.