Retry open file operations when attempting to save source files / source database files.
When dropbox or other file syncing solutions are present and attempting to sync source files that are active in the editor, save operations on said files can fail due to the inability to open the file (as it is being read in exclusive mode by the backup software).

This change makes it so that now we retry these operations up to a certain amount of seconds, specified by a new user preference save_retry_timeout.