Group LaTeX scape sequences (generally)
In general, it is better to group these escape replacements in order to ensure they render properly when inlined inside of other strings. In testing, I ran across a handful that need to be ungrouped (as they actually apply hats or other modifications to following characters). This change default encloses output LaTeX in ‘{}’ unless specified by the character definition.