Fix incorrect status message when searching crossref
The selectedPanelProvider wasn’t being passed to a memoized callback, which means that the provider was captured at the time the callback was memoized. Instead, pass the provider so it reflects the current state.

(in this case, what this meant was that the wrong the provider was being queried for typeahead searches/status, though the bug didn’t appear that bad b/c there was only one typeahead provider).