Support Insert Citation inside Existing Citation
When we insert the citation, this will now check whether a cite mark is already active. If it is, we will skip inserting the wrapper (e.g. `[` and `]`) and cite mark. Note that we skip inserting the wrapper even if the user requests it since it is nonsensical to nest cite marks in that way.