Always report startup configuration related errors to stderr, even if not connected to a terminal.
This is done because when installing on CentOS, we want to see these error
messages. However, due to how that process works, the terminal is not actually
connected to the installing rserver process. Any errors encountered will be logged
to syslog, but they will not directly be shown to the user. Forcing these writes
to stderr ensures they show up.

Fixes https://github.com/rstudio/rstudio-pro/issues/2214.