first commit for async rpc - always on or off based on config
When accepting a new http connection, check if it can be made async.
If so, respond to http with the asyncHandle and convert the HttpConnectionImpl
into an AsyncRpcConnection (this involves parsing the JSON)

- new virtual method for sending the json response:
   - for normal http, it sends the http response
   - for async rpc, it emits the kAsyncCompletion event

for now, there's a global session-aync-rpc-enabled configuration
option set to 1/0 to turn it on/off. It's currently off by default.

Async rpc is disabled until we see the first get_events
request (since the client doesn't listen for kAsyncCompletion events
until then)