Fix problems with Windows implementation of FileLogging.
The file must be opened and closed for every log operation. Previously, this was broken such that subsequent log invocations would fail to write to file on Windows since the file cannot be opened more than once at a time.

Also shortened the run duration of the tests because on Windows the filesystem operations involved a pretty slow so the unit tests take quite a long time to run.