use 'env' as the standard python project environment directory rather than '.venv'
I've been hashing out the project-specific virtual envs a bit and thinking more about directory naming. I noticed that VS Code will actually auto detect/activate any venv contained in a sub-directory with a pyvenv.cfg file. That's great flexibility, but at the same time I think we want to enforce some standards so that the IDE can depend on a predictable path (e.g. we need to screen files in the venv out of find in files results -- we  do that now for "renv" but need the analogous behavior for python envs; we also need to augment .gitignore with the venv directory).

I noted that the official python venv documentation (https://packaging.python.org/guides/installing-using-pip-and-virtual-environments/#creating-a-virtual-environment) use the convention of 'env' for project-specific environments.

I also noted that this article from the Software Carpentries (https://carpentries-incubator.github.io/introduction-to-conda-for-data-scientists/02-working-with-environments/index.html#conda-environment-sub-directory-naming-convention) recommends env for compatibility with common venv practice as well as to get the automatic .gitignore from the standard GitHub python .gitignore file. This article on using conda environments (by the same author) is highly ranked on google and recommends the same thing: https://towardsdatascience.com/managing-project-specific-environments-with-conda-406365a539ab

We could adopt VS Code's approach and be more aggressive about auto-detection (as well as make our find in files and .gitignore code similarly flexible). I think though that the community is better served by the adoption of a common, sensible convention that is based on official documentaiton/recommendations (it's nice that we can point users to the above mentioned articles and know that if they follow them precisely the right thing will happen in the IDE).

We can certainly make this more flexible if we get feedback that this is desirable, but I vote that we start w/ a simpler approach.