call finishInterrupt only when session has really been interrupted
With async rpc, the interrupt requests are responded to right away
even if the session is still running. That broke the feature that
allows you to click interrupt three times to get the abort dialog.

With this PR, the busy status only processes the interrupt as finished when
busy is set to false