Disable Caching When Zotero is Running Entirely Locally
It appears that the version counter that is present in a Zotero database is only incremented when Zotero sync is configured. In the case of a completely unlinked / unsynced local database, the version remains 0 at all times. As a result, we would check versions and cache the collection even as it changes. This fix causes our caches to no longer effect collections whose versions are 0.

Fixes #7722