Fix logging bug where chowning the logfile before writing to it does not work.
The following sequence of events will prevent the log file from being chowned:

1. Log file was previously created as root and was never chowned to service user.
2. Process starts up again at a later time and goes to drop privilege thus attempting to chown the log.
3. Log cannot be chowned because the LogFile variable is not properly initialized.

This is because initialization currently only occurs when the log file is attempted to be written to.

This change also makes sure that the initialization happens in the constructor as well as during
calls to write, to ensure that the path of the log file is always properly initializated and
up-to-date.