Fix bug where incorrect parameter was being specified when constructing a FileLogOptions.
This caused the FileLogOptions(directory, warnSyslog) constructor to
always cause the maximum number of rotations to be 1 instead of 100.
This was preventing older important Launcher log data from being shown
when enable-debug-logging=1 was set.

In terms of IDE impact, this bug causes the monitor log file to only
rotate once, instead of a maximum of 100 times (keeping old log files
around for 30 days).