Clear the cite key anytime the 'x' is clicked
When the user selects a citation in the insert cite dialog, we automatically add it to the list of citations to insert. Make the 'x' icon in the tag function in this case by removing the selection (since the fact that the item was selected is the reason that the item is being included in the add list).