Fix for Verify Installation
RActiveSessionStorage was making assumptions about the context when a
ActiveSessionStorage was created, using our XDG to discover the storage
directory.

Launcher creates some sessions to do verification but the XDG is not
populated, so it was searching in the / of the filesystem.

Removing unused or easily substitutable constructors for ActiveSession
to increase clarity.

Switched to the FileActiveSessionStorage always taking a directory,
which will be the scratch path (session-<id>) path. Gets
FileActiveStorage out of the storage discovery game.

Removed unused factory methods, we can reintroduce them later, but they were leading us astray here.
The id will still be necessary for DB calls to read and write properties