Adds the ability to passthrough log messages and add structured data to log messages.
The passthrough logging is used to enable logging streams from child processes to seamlessly
be combined into the parent's log stream. This will be primarily used for the Launcher and its
plugins, but could also be used in RSW as well for smaller utility binaries to log to the main
rserver.log file.

Also adds the ability to add LogMessageProperties to a logging entry, which can contain
arbitrary structured data. This allows us to use a structured logging approach, which
primarily has benefits when used with JSON log format and log search tools like Kibana.