Avoid offline requests before the session is initialized (#10347)
This fixes a problem that showed up when running the test suite
with session-handle-offline-timeout-ms=0 - basically, running all
offlineable requests on the offline thread. Every once and a while it
would try to run an offline request before the session was initialized.
That would initialize the session on the wrong thread, getting a ton of
errors and eventually a crash.

I don't think it's likely without that setting but just to
be safe I also added the session init test to the offline thread as well.