Modify library code to synchronize with Launcher repo.
There are some spots that are different between the codebases because
of the different log paths, so made those into a compile time variable.

Also, the Launcher does not use the database code (or SOCI at all), and
cannot compile it currently, so it is excluded depending on a compile
time variable.