Add 'starting' debug log right after log is initialized
Addresses the remaining aspect of: https://github.com/rstudio/rstudio-pro/issues/2641

We already have debug logs beginning the R session, before and after resuming the session, and when
listening. This adds a message right when the session starts, and prints the dyn library path value
for linux and mac.

Should I log PATH for windows?