For non-launcher sessions, need to pass the server rpc key
Initialize the socket rpc before the session scope is initialized
in the options so the initial "validate" works.

We already pass the server's rpc secret key for OS/Server since it's in the server's
environment and the fork/exec inherits it but we don't for non-launcher
sessions so this adds that.