Remove the redundant libedit2 dependency for Debian packages.
This was originally added in 2e7e7ab03c797f39d64097090ca3cc4fe8e710f8 to
support the bundled libclang (since libedit is a dependency of LLVM).
However, since 543edee3924e843fdad936e2d19e7c2efe088d38 we depend on the
system libclang-dev on Debian-based distros, which was part of the move
away from using the bundled libclang on Linux (#2892).

Since libedit2 is a dependency of libclang-dev on all supported
Debian-based distros, we can safely remove this explicit dependency.

To verify that libedit2 is a transitive dependency of libclang-dev, you
can check the following:

    $ apt-cache depends --installed --recurse -i libclang-dev | \
        grep 'Depends: libedit2'

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>