Make libpam optional in the core libraries.
Launcher does not use any libpam features from server_core and it
complicates the build and CI process to require it unconditionally. This
commit allows Launcher to opt out of using libpam (and in fact that is
the default).

This conditional include is similar to the one carved out for the
database/SOCI libraries, which Launcher *also* does not use.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>