Add support for session memory limits & usage with cgroups v2.
This commit adds detection of memory limits and usage when the Linux OS
uses cgroups v2. This is the default in Ubuntu 22.04 and Debian 11.

The commit is structured to make minimal changes to the existing
MemoryProvider design, but for convenience I've adopted LONG_MAX as a
sigil value because cgroups v1 and v2 use different representations.
This makes some of the existing code simpler.

To run the unit tests in a cgroups v2 environment with a memory limit,
try something like the following:

    $ systemd-run --user -p MemoryHigh=100M --pipe --wait build/src/cpp/core/rstudio-core-tests Resources
    $ systemd-run --user -p MemoryMax=100M --pipe --wait build/src/cpp/core/rstudio-core-tests Resources

This partially addresses #10914, but does not fix the monitor-specific
portions of that issue.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>