Introduce rpostback-browser
Some software uses the BROWSER environment variable to
launch web browser windows (notably, both R's browseURL
and Python's webbrowser.open()). In Terminals and
subprocesses of RStudio Server sessions, any attempt to
launch a browser would fail.

This commit introduces rpostback-browser, which takes a
single command line argument that is a URL to open; this
causes the server to translateLocalUri on the URL if
necessary, then ask the current session to launch it in a
new window. It also sets the BROWSER environment var
to rpostback-browser from within rsession, where it will
be inherited by the Terminal and any subprocesses.