Change the command used to run PyShiny apps
Since the modified code was written, two new features
were introduced: the ability to launch a browser, and to
select a random port upon startup. These features are
both useful for RStudio Server: launching a browser (via
$BROWSER) frees the server from having to know when
the app is ready for action, and the random port makes
it possible for multiple users to run independent PyShiny
apps on the same server.