Allow localhost-port-proxying feature to work with chunked encoding
When localhost HTTP endpoints used chunked encoding and we
attempted to proxy them through /p/<token> URLs,  they were
failing with "invalid chunk encoding" or something. This is because
during the course of localhost proxying, the LocalhostAsyncClient
was decoding the chunked encoding, so the response's body was
actually just raw data, yet the "Transfer-Encoding: chunked"
response header remained. This commit detects this condition
and changes the headers to reflect a normal non-chunked
response.

I'm not sure what happens if the upstream response is not only
chunked but also, say, gzipped (and this can happen in either
order: "gzipped, chunked" or "chunked, gzipped".