Fix RSA signature buffer initialisation for old versions of OpenSSL.
This commit initialises the siglen parameter as required by the OpenSSL
docs for EVP_PKEY_sign() [0]:

> If sig is not NULL then before the call the siglen parameter should
> contain the length of the sig buffer, if the call is successful the
> signature is written to sig and the amount of data written to siglen.

The previous code worked against the OpenSSL 3 library I QA'd it against
and in unit tests (which have small payloads), but Cole Arendt reported
that it failed in real-world usage on a Ubuntu 18.04 environment:

    2022-08-30T15:23:32.112970Z [rserver] ERROR system error 74 (Bad message) [description: error:0608C09B:digital envelope routines:EVP_PKEY_sign:buffer too small, description: Error signing Launcher session request, request-uri: /events/get_events]; OCCURRED AT rstudio::core::Error rstudio::core::system::crypto::rsaSign(const string&, const string&, std::__cxx11::string*) src/cpp/core/system/Crypto.cpp:169; LOGGED FROM: void rstudio::server::session_proxy::{anonymous}::logIfNotConnectionTerminated(const rstudio::core::Error&, const rstudio::core::http::Request&) src/cpp/server/session/ServerSessionProxy.cpp:503

I have not precisely identified the root cause, but I notice that
OpenSSL changed how this parameter is handled internally in 2021 [1].
That OpenSSL commit in question describes the fix as requring an
"application bug" to hit the issue and cites no CVE issued -- so it's
likely that not distros will backport the patch and we should expect
different behaviour among them.

Since the previous code was, according to the documentation, an
"application bug", it seems likely that this change will resolve the
issue for everyone.

[0]: https://www.openssl.org/docs/manmaster/man3/EVP_PKEY_sign.html#DESCRIPTION
[1]: https://github.com/openssl/openssl/commit/43da9a14f0e73f42f28ae34219929b44df5d1a11

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>