Fix WebSocket proxy corruption
When proxying Shiny for Python traffic, occasionally the
WebSocket proxying would fail; as soon as the WS was
opened, it would close. This was because the writing of
the WS response (e.g. the HTTP headers being returned
from the server to the client) was overlapping with the
writing of the WS frames that followed, i.e. we weren't
waiting for the HTTP header writing to complete before
doing more writes.

The changes in this commit let us pass a callback to
AsyncConnection.writeResponse(), and we wait to
proxy WS traffic until that callback is called.

Separately, I noticed by inspection that streaming
responses in AsyncConnectionImpl didn't respect
the writeResponse method's 'close' argument; I
tried to fix this but I wasn't able to actually try my
new code, I don't know how to get RStudio to
send such a response.