handle SigTerm just like SigUsr2
The goal here is to work more nicely with supervisors like `systemd`
without requiring a custom shutdown script or other unit customization.
It is possible that ignoring SigTerm on Desktop alters some other
implicit functionality, but we thought this would be preferable and more
explicit.

open source companion to rstudio/rstudio-pro#3811