Fixed bug where failing to delete a rotated log file could cause a log line to be emitted, causing deadlock.
This is because the call to FilePath::getLastWriteTime can fail and
end up emitting a log line, which causes recursion in the logging
code leading to attempting to re-acquire a mutex that was already
acquired.

The fix is to make sure we aren't logging by adding a new method
overload that returns the error instead of logging it.