Avoid non-zero exit codes when running sessions under Launcher.
Users sometimes raise concerns about the number of nonzero exit codes
they see from Workbench sessions on Slurm and Kubernetes [0]. This is
because we use exit code 100 and 101 to signify switching between
projects and suspending, respectively, both of which are common
operations in these environments.

Another side effect of this is that many pods in Kubernetes are marked
as being in a somewhat-alarming "Error" state when finished rather than
the benign "Completed" state (see
https://github.com/rstudio/rstudio-pro/issues/3375 for details).

It seems that we don't rely on these exit codes when running sessions
under Launcher at all, so this is an easy change to make.

In my manual testing, switching between projects and suspending sessions
continues to work just fine with zero as the exit code.

[0]: For a recent example, see:
     https://rstudio.slack.com/archives/C7312LBPY/p1670494955085729

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>