Stub out a process creation time lookup for macOS.
We have an existing utility to look up the creation time of a process,
but the only implementation is for Linux. This has been fine so far, as
Launcher seems to be the only consumer of this API, but the compile-time
guard is now generating issues for us when trying to build that project
on macOS.

So this commit stubs out a macOS implementation that returns an error.
Ideally, we'd actually implement this for macOS, but doing so turns out
to be nontrivial [0] so I don't think it's worth pursing at this time.

[0]: https://stackoverflow.com/questions/31603885/get-process-creation-date-time-in-osx-with-c-c

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>