Lower CPU priority during session suspend.
Large sessions can be CPU-intensive to serialize, and customers have
often reported mysterious CPU spikes that were tracked down to session
suspension. This CPU spike can be particularly inconvenient when it is
caused by previously-dormant sessions that time out and are
auto-suspended by Workbench on a busy system.

Recently, a customer asked [0] if we could lower the session process's
nice value right before suspending it. This seemed like a good idea to
me, so this commit implements this for macOS and Linux. We could also do
it on Windows with SetPriorityClass(), but on that OS the likelihood of
being a multi-user system is much lower.

This change doesn't eliminate the CPU required to suspend, of course,
(or the I/O to do so, which can be significant) but it does make
suspension much less likely to interfere with other, more important
workloads on the system.

[0]: https://app.prodpad.com/feedback/874d975e-9b8c-11ed-a544-2a7db0eb1d9c/canvas

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>