Enable WAL mode and set a busy timeout for SQLite connections.
Right now we don't have much load on the database, but recent
scalability testing surfaced many "database is locked" errors from the
login flow in the logs. In some cases these errors are merely advisory,
but in others they propagate up into issues in the UI.

Currently, we run the SQLite database in the default "rollback journal
mode". The most common advice for addressing "database is locked" issues
is to instead enable WAL mode, which allows readers and writers to
operate concurrently, and offers much better performance in general.

So this commit enables WAL mode on our database.

However, note that locking issue are not completely eliminated. From
SQLite's documentation on WAL mode [0]:

> [An advantage] of WAL-mode is that writers do not block readers and
> readers to do not block writers. This is *mostly* true. But there are
> some obscure cases where a query against a WAL-mode database can
> return SQLITE_BUSY, so applications should be prepared for that
> happenstance.

As a guard against these cases, this commit also sets the SQLite busy
timeout [1]. This five-second timeout (and the use of WAL mode) match
what is being used successfully by Connect, and should dramatically
reduce the incidence of these "database is locked" errors.

Finally: it's still *possible* to encounter SQLITE_LOCKED or SQLITE_BUSY
errors with these changes, but the only remedy in those cases is to
catch the SQLite-specific exceptions from SOCI and retry the query using
our own backoff logic. This commit does not attempt to add this level of
complexity to our query implementation, since the appropriate amount of
time waiting for a query is pretty dependent on the context it's being
executed in, and I'd rather see this responsibility pushed up to the
caller.

[0]: https://sqlite.org/wal.html#sometimes_queries_return_sqlite_busy_in_wal_mode
[1]: https://sqlite.org/c3ref/busy_timeout.html

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>