Add support for serving static files precompressed with Gzip/Brotli.
At present, we compress static files at runtime to reduce the bandwidth
cost of sending them over the network. Unfortunately, this compression
is by no means free, and we incur significant CPU overhead to serve
files this way (not to mention the latency added while we do so).

This has an especially large performance impact for the large
Javascript/CSS assets on the Workbench Homepage, which are static files
that never change.

It's common practice to avoid the runtime compression cost for static
assets by precompressing the files beforehand, either with Gzip or, more
commonly these days, with Brotli.

This commit modifies the static file handler to substitute in
precompressed files when they exist, much like NGINX's popular
gzip_static and brotli_static directives.

Note that this patch *does not* actually compress any of the assets that
we serve. It only makes it possible to do so.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>