remove `openExternal` call from `allowNavigation` check
`allowNavigation()` checks if navigation should be allowed, but in the final `else` case, it also opens the specified url via `shell.openExternal()`.

This side effect of opening the url should be removed from the function because the function should only serve a single purpose: to check if navigation should be allowed.

Other code which calls `allowNavigation()` uses the boolean result to determine if the url should be opened or not, which seems to be why multiple opens are getting triggered (`allowNavigation()` may open a url and the calling code may open the url again).