Send SIGKILL to child processes during forced exit to avoid zombies.
The Package Manager team initially noticed this when the Kubernetes
Launcher plugin would hang indefinitely but never exit. Jeff has also
found the symptoms of this in some customer logs.

We already "force" an exit event after 30 seconds, this commit just
piggybacks on that timer to send it SIGKILL at that point as well.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>